My Integer Lab
Design a class named MyInteger. The class contains:

A private int data field named value that stores the integer value represented by this object.
A constructor that creates a MyInteger object with given value. If no value is provided a default of 0 is used.
An int getValue() method that returns the value.
A void setValue(value) method that updates the this.value.
Boolean methods isEven(), isOdd(), and isPrime() return true if the value is even, odd, or prime, respectively.
Static boolean methods isEven(int value), isOdd(int value), and isPrime(int value) that return true if the specified value is even, odd, or prime, respectively.
Static boolean methods isEven(MyInteger value), isOdd(MyInteger value), and isPrime(MyInteger value) that return true if the specified value is even, odd, or prime, respectively.
Boolean methods equals(int value) and equals(MyInteger value) that return true if the this.value in the object is equal to the specified value.
A static method int parseInt(char[] value) that converts a value array of numeric characters to an integer. For instance, parseInt( ['1', '2', '3'] ) would return 123. Do not use the Integer class.
A static method int parseInt(String value) that converts a string value into an integer.