Roman Number Homework
Design a class named RomanNumber. The class contains:

A private int data field that stores the value represented by this object.
A private String data field representing the value in Roman Numbers. 
Constructors that create a RomanNumber object for the Roman Number specified as a String or an int.
A getNumber method that returns a String representing the Roman Number.
A getValue method that returns an int representing the number.
A static method parseValue(String num) that converts a String representing a Roman Number into an int.
A static method parseNumber(int num) that converts an int num into a String representing the Roman Number .
Methods equals(String) and equals(int) that return true if the value in the object is equal to the specified value.
Methods add(String num), subtract(String num), multiply(String num), divide(String num),  that return a String representing the Roman Number after addition, subtraction, multiplication, integer division, respectively, of the object value with the num.
A static method calculate(String expression) that performs calculation of a String of two roman numbers separated with an operand, and returns a String representing the resulting Roman Number. For instance, you may get RomanNumber.calculate(“Ⅿ/Ⅺ”) or RomanNumber.calculate(“ⅭⅬⅦ-ⅬⅩⅢ”). 
Methods min(String num) and max(String num) that return a string representing the maximum and minimum, respectively, of the value of the object and given num.