Triangle Class Lab
Design a class named Triangle that extends GeometricObject Download GeometricObject. The Triangle class should contain:

Three double data fields named side1, side2, and side3 to denote three sides of the triangle.
A no-arg constructor that creates a triangle with default values 0.0 for each side.
A constructor that creates a triangle with the specified side1, side2, and side3.
The accessor methods (i.e., set & get) for all three data fields.
A method named getArea() that returns the area of this triangle.
A method named getPerimeter() that returns the perimeter of this triangle.
A method named isEquilateral() that returns whether all sides are equal.
A method named isIsosceles() that returns whether two sides are equal. Note that an equilateral triangle is also an isosceles.
A method named isScalene() that returns whether no sides are equal.
A method named toString() that returns a string description for the triangle by appending to the parent class method.