Rectangle Class Lab
In this assignment, we are building on the Lab 7 - Triangle Class to design a class named Rectangle that also extends GeometricObject Download GeometricObject.

The Rectangle class should contain:

Two double data fields named side1 and side2 to denote opposing sides of the rectangle.
A no-arg constructor that creates a rectangle with default values 0.0 for each side.
A constructor that creates a rectangle with the specified side1 and side2.
The accessor methods (i.e., set & get) for all data fields.
A method named getArea() that returns the area of this rectangle.
A method named getPerimeter() that returns the perimeter of this rectangle.
A method named isSquare() that returns whether both sides are equal.
A method named equals(Rectangle rectangle) that returns whether this instance is equal to another rectangle instance. Note that, two instances are equal only if side1 and side2 of both instances are equal.
A method named toString() that returns a string description for the rectangle by appending to the parent class method.
Modify the GeometricObject to implement the Comparable interface using the areas of geometric objects. (see https://liveexample.pearsoncmg.com/html/ComparableRectangle.htmlLinks to an external site. as example)

You would need to declare a compareTo method and an abstract getArea method, which is implemented in both subclasses.