Directory Report Lab
Design a utility class named DirectoryUtil.

Your main method should produce a file named "[folderName].log" report for the given folderName, which is obtained as an argument.
The report should produce the number of files and file sizes within the folder. 
Each line of report would be "fileCount folderSize path" where each is separated by a tab. Note that fileCount is only for files (not folders), and folderSize is in bytes and includes the subfolders of the reported folder.
For example, in the following output courses folder includes both ssw315 and ssw810 folder's counts and file sizes.
You may use a stack to record subfolder's reports so that log file has the correct ordering of parent folders being reported before their subfolders.
You may use file.getPath() method to obtain full path of a file.

Score: 8/10