Code Validator Lab
Design a utility class named CodeValidator that would verify the grouping match of a given code. Given a java source code, verify that the characters ", ', (, ), {, }, [ and ] match.

Open brackets must be closed by the same type of brackets.
Open brackets must be closed in the correct order.
Every close bracket has a corresponding open bracket of the same type.
Contents of a string "..." and character '...' are ignored. For simplicity, assume there are no special characters in a string or char. That is, every " and ' is a beginning or end of a string and character, respectively.
Contents of comments after // are ignored. For simplicity, assume there is no multi-line comment with  /* ... */ 

Score: 10/10