Baby Names Homework
Write a program that produces a ranked list of the top ranked baby names per year from the IRS dataLinks to an external site. (do not use other data sources). 

You would need to aggregate the state-level data into two sets of tables for girls and boys.
The user would provide the base year (i.e., for which the ranking would be based) via the command line argument.
If user indicates "all" as the base year, then the total counts across all years would be used to produce the rank.
If user indicates "YYYYs" as the base year where YYYY is the year ending in a '0', then the total counts for the decade would be used to produce a rank.
Produce a comma-separated CSV file that tabulates the top 100 names for the selected year along with those names' rankings in the other years. 
First row should be:  name, 1910, 1911, ... 2020, 2021
After the first row, each row indicates the ranking of a name (sorted by the base-year ranking) for each of the years.
After the first column, each column indicates the ranking of names for the particular year. Years should be sorted from the lowest to the highest.
Note that the selected year's rankings will increase from 1 to 100, but other years would be based on the names column.
Do not use a specialized CSV library.

Score: 9.95/10